add an implementation of computing longest common subseuquence length
use a dynamic programming matrix  to compute the length. This is much much faster than the implementation that relies on substringing etc.
also added a test that constructs a really long string and ensures the right value is returned. The test ran on my machine in under a second whereas the recursive implementation doesnt return for any realisitic amount of time.